cleanup and removing old code
