+++ templates/01_load_secb_data.py
@@ -3,8 +3,7 @@
 from pathlib import Path
 import numpy as np
 from pyhdx import PeptideMasterTable, read_dynamx, HDXMeasurement
-from pyhdx.process import apply_control, correct_d_uptake
-from pyhdx.support import filter_peptides
+from pyhdx.process import apply_control, correct_d_uptake, filter_peptides
 
 current_dir = Path(__file__).parent
 output_dir = current_dir / "output"
