+++ tests/test_fileIO.py
@@ -16,8 +16,7 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from pyhdx.process import apply_control, correct_d_uptake
-from pyhdx.support import filter_peptides
+from pyhdx.process import apply_control, correct_d_uptake, filter_peptides
 
 cwd = Path(__file__).parent
 input_dir = cwd / "test_data" / "input"
