+++ tests/test_models.py
@@ -13,8 +13,7 @@ import tempfile
 import pickle
 import pytest
 
-from pyhdx.process import apply_control, correct_d_uptake
-from pyhdx.support import filter_peptides
+from pyhdx.process import apply_control, correct_d_uptake, filter_peptides
 
 cwd = Path(__file__).parent
 input_dir = cwd / "test_data" / "input"
